requirements.txt
# TradingView Agent - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Broker APIs
python-binance==1.0.19          # Binance (Spot, Futures, Margin)
oandapyV20==0.7.2               # OANDA Forex/Gold
-e ./external/coinbase-advanced-py  # Coinbase Advanced Trade SDK (vendored)

# Web3 / Base DEX
web3==6.11.3                    # Ethereum/Base chain interaction
eth-account==0.10.0             # Wallet management
eth-abi==4.2.1                  # ABI encoding/decoding

# Google Cloud
google-cloud-secret-manager==2.17.0
google-cloud-logging==3.8.0
google-cloud-firestore==2.14.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
pyyaml==6.0.1                      # YAML configuration
yfinance==0.2.28                   # Yahoo Finance for price fallback

# Monitoring & Alerts
python-telegram-bot==20.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security
cryptography>=41.0.0  # Let pip resolve compatible version

